+++ crates/fs/src/nova/algorithms/mod.rs
@@ -1,0 +1,4 @@
+pub mod preprocessor;
+pub mod key_generator;
+pub mod prover;
+pub mod verifier;
