+++ crates/fs/src/nova/circuits/mod.rs
@@ -1,0 +1,1 @@
+pub mod verifier;
